move store to package core/store
